requirements.txt
pytest
Pillow
appium-python-client
py
dominate
pytest-ordering
pytest-html 